@@ -2,10 +2,9 @@ import { shallowMount } from "@vue/test-utils";
 import CopyButton from "@/components/CopyButton.vue";
 
 describe("CopyButton.vue", () => {
-  it("renders", () => {
-    const data = "OK";
+  it("renders props.data when passed", () => {
     const wrapper = shallowMount(CopyButton, {
-      propsData: { data }
+      props: { data: "OK" }
     });
     expect(wrapper.text()).toMatch("OK");
   });
